COMP: Remove undefined complex types
http://www.cplusplus.com/reference/complex/complex/

Type of both the real and imaginary components of the complex number.  The
effect of instantiating a complex with a T other than float, double or long
double is undefined (certain library implementations may support it, but the
resulting code is non-portable).

http://stackoverflow.com/questions/11108587/why-does-abscomplexint-always-return-zero

According to the C++ ISO spec, §26.2/2:

The effect of instantiating the template complex for any type other than float,
double or long double is unspecified.

Change-Id: I470b22f8e50ac0206f20646e342f19f6ea127f14