BUG: TvirtualImage mismatch in registrationV4 and metric object
Consider a situation that we want to register two float images using a
double type transform.
We instantiate our metric as an object of itkImageToImageMetricv4 class
or one of its derived classes.
In the definition of the metric classes, the TvirtualImage is set to TFixedImage,
and the InternalComputationValueType is set to double by default.
Therefore, in the case of our example, the metric object has the type of:

  itkImageToImageMetricv4< Image<float,Dim>,
                           Image<float,Dim>,
                           Image<float,Dim>,
                           double>                  (1)

The above metric object is set to the itkImageRegistrationMethodv4
class as its base class type:

  itkObjectToObjectMetricBaseTemplate<double>

because the metric type in ITKv4 registration framework is defined as:

  typedef itkObjectToObjectMetricBaseTemplate<RealType> MetricType;

Where the RealType comes from the output transform type:

  typedef typename OutputTransformType::ScalarType RealType;

That is "double" in our example.

Inside the registration framework, the above MetricType object should be
casted to the ImageMetricType that is defined as:

  itkImageToImageMetricv4< Image<float,Dim>,
                           Image<float,Dim>,
                           Image<double,Dim>,
                           double>                  (2)

As in ITKv4 registration method class, we have:

  typedef ImageToImageMetricv4<FixedImageType, MovingImageType,
                              VirtualImageType, RealType> ImageMetricType;

Where the virtual image type is defined as:

  typedef Image<typename ObjectToObjectMetricBaseTemplate<RealType>::ParametersValueType,
               ImageDimension>   VirtualImageType;

That in our example, the RealType is double and the virtualImageType
have the type of: Image<double,Dim>.

Obviously a dynamic cast from (1) to (2) will be failed and causes a
segfault in current version of ITK.

To avoid this issue, we should have consistent definitions for the
virtualImageType in the metric classes and RegistrationMethodv4 class.
Therefore, we should be able to pass the VirtualImageType to
the ITKv4 registration framework as a template parameter that is set to
TFixedIamge by default the same as metric classes.

A new test is added to ITKv4 to show above bug. This test returns
failure with current version of ITK; however, it will be passed after
applying the changes introduced by this patch.

Change-Id: I06e6f8fcee8e7e54f23e0e05054be77495f82c68