ENH: Improved writing speed for TIFF images
Previously, in itkTIFFImageIO.cxx rowsperstrip was set to a default
value so that it would be calculated using the STRIP_SIZE_DEFAULT
defined to be 8 kB in tiffiop.h. However, this a very conservative
small number, and it leads to very small strips resulting in many io
operations, which can be slow when written over networks that require
encryption/decryption of each packet (such as sshfs). Conversely, if
the value is too high, a lot of extra memory is required to store the
strips before they are written out. Experiments writing TIFF images to
drives mapped by sshfs showed that a good tradeoff is achieved when
the STRIP_SIZE_DEFAULT is increased to 1 MB. Rather than change that
value in the third party libtiff library, we instead compute the
rowsperstrip in itkTIFFImageIO.cxx to lead to this same value.

This results in an increase in memory usage but no increase in writing
time when writing locally and significant writing time improvement
when writing over sshfs.  For example, writing a 2048x2048 uint16
image with 8 kB per strip leads to 2 rows per strip and takes about
120 seconds writing over sshfs. Using 1 MB per strip leads to 256 rows
per strip, which takes only 4 seconds to write over sshfs.

This has no effect on accuracy, only writing time.  All of the tests
still pass.

Change-Id: Id3a99fbcbcd430e6436dbf262aab2fc70b82937a