DOC: Better description of what the code is doing.
Improved the code organization and variables names in order to better convey
the actions that the code is taken. This was driven by a warning reported by
a Coverity scan with text:

  1103094 Out-of-bounds access
  Memory not owned by this buffer will be accessed, causing memory corruption
  or incorrect computations.

  In

      itk::​Functor::​ScalarToRGBPixelFunctor<unsigned char>::
      ​operator ()(unsigned char const&):

  Access of memory past the end of a memory buffer

The scan warns about the risk of using out of bounds elements due to the use
of an array/pointer to navigate through the bytes of the scalar class. This
is however, the correct intent of the code, and bounds are checked properly
using the sizeof() the scalar type.

Change-Id: I1884ca33f5a7eea8bb28b043f8bcef4bff129805