ENH: Update CMake deprecated commands with new equivalents
Note that "add_subdirectory" command is evaluated immediately, while
"subdirs" is evaluated after the current directory's CMakeLists (thanks
Brad King for pointing this out). For the Ruby and Tcl wrapping, the
"Tests/CMakeLists.txt" file does not rely on any of the work that the
upper-level CMakeLists does between the call to "add_subdirectory" and
the end of the CMakeLists file (namely defining 2 macros).

The "itkExtras" subdirectory does not exist in the Ruby wrapping, so
the line referencing it was removed (Ruby wrapping is still a WIP).
Change-Id: I5d23aa06607ed243c18cb1ec5eacedf4092c33c5