BUG: DCMTKFileReader erroneously propogating exceptions
The GetElement... methods have an optional 'throwException'
flag. The purpose is to give the caller a choice of handling
a failed call locally (by e.g. looking for an alternate DICOM
tag for the same data) or for an exception to be thrown in
the case where there's no recovery from a failed GetElement.

GetElementFLorOB didn't propagate the throwException flag down
into the GetelementOB function called if GetElementFL failed,
which meant an exception was getting thrown when the caller
requested no exception.

Change-Id: Iccb6ebe90e6a62c238f9cbbe89fd8224bbca0161