COMP: Fix gcc_xml segfault due to structure alignment
This patch addressed the following compilation issue with WrapITK for
gcc_xml:
itkSize.h:53: internal compiler error: Segmentation fault
Please submit a full bug report,

The way the alignment attribute is currently used in ITK is known to
cause segmentation faults in older gcc compilers. There are try
compiles for this. However, this is also happening for gcc_xml in this
case. This patch is really a hack. WrapITK does not need to know about
these private data structures, so we use the processor to guard the
problematic section for gcc_xml.

Change-Id: Ic24342ac7dc45e6687390c0eee760d081be2cc2a