BUG: SpacePrecision typedef's incorrect.
In a few places the precision of points or continuous index
were allowed to be non-double. In those cases it is critical
that the previous typedef's be maintained.  This patch reviewed
where typedef's were changed from previous specializations to the
global SpacingPrecisionType incorrectly.

Tested external builds for ANTS, BRAINSTools, and the internal
test suite to ensure that this reversion fixes the regressions.

Change-Id: I2f94e0601e84230c1001af02a90799270df2b582