PERF: Simplify conditionals in loop
This is to simplifying by removing unnecessary conditionals
to provides additional optimization opportunties.  The conditionals do not
change the values, but they do attempt to reduce the amount of computation
that needs to be done.  In reducing the amount of computation, it also
increases the complexity of the code and introduces alternate code paths
that make optimization more difficult.

        187.2   175.58
        189.42  177.67
        188.91  178.78
        190.58  183.05
        191.37  191.53
Average 189.496 181.322 4.508%
Stddev    1.605   6.323

Change-Id: I92e98ba8b44b8891c1f3277565d6d51e3b7806a3