BUG: Fix remote image IO module registration.
When an external application uses a build of ITK that has
a remote IO module turned ON, the expected behavior is that
the corresponding IO factory should be regisitered automatically
like any other internal image IO types.

MGH and SCIFIO (currently the only two remote IO modules) IO
factories are not registered when their modules are enabled, due to
the module naming convention difference (with the internal IO modules):
there is no "ITKIO" prefix in the remote IO modules' names.

Change-Id: I2c6e3ef8606d2337f6dc5593267f4c26d134aecb