BUG: Remove an invalid test case that will always fail
The test itkImageReadDICOMSeriesWriteTest takes an input of a MetaImage,
writes it to a DICOM series, re-reads the DICOM series, and compares the
re-read image to the original. However, the particular MetaImage
"HeadMRVolumeWithDirection003.mhd" uses a left-handed coordinate system,
which is incompatible with DICOM (DICOM defines its coordinate system to
be LPS, which is right-handed).

The test case itkImageReadDICOMSeriesWriteWithDirection003Test used
"HeadMRVolumeWithDirection003.mhd" as its input data, so it would always
fail. Better error handling when attempting to write left-handed images
to DICOM is needed, but for now, the test case will be removed.

The file "HeadMRVolumeWithDirection003.mhd" was not removed or modified,
as is still used for the case itkImageIODirection3DTest09. Test coverage
for the ImageBase module is significantly more important than the
Deprecated module.

Change-Id: I109edf2e2e0a7fd971cdc2b2229e15781a38ab11