BUG: Add new baseline and relax tolerance for histogram plotter
A new baseline was generated from gcc 4.4.7 on RH5 for i686, which
error appears to match that of gcc 4.8.2 on Debian
32-bit. Additionally this demonstrates that it's not a compiler issue
but a difference architecture numerics.

The tolerances for the comparison are also significantly relaxed due
to the sensitivity of the join histogram. Specifically, a large radius
is used since the intensity of the registration is mapped to location
in the plot. Also the goal it to verify the histogram is correctly
plotted not evaluate the accuracy of the registration.

Change-Id: I1937266dc4edaebc1383ed28916da86e130e4067