BUG: Fix for the WrapITK.pth destination path
Added a regex to remove the \n character at the end of the site_packages path found by cmake.
On OS X this would result in a wrong path for the WrapITK.pth file, so that the python linking was broken.
Note that I was not able to find out what changed since ITK 4.4.2, perhaps a change in cmake or python 2.7.6 ?

Change-Id: Ic5f985538058e625aea46d936fe0aa55b7d1ba2e