BUG: Adding export specification to Exception objects
The ProcessAbort exception was not being handled correcly accross
compilation units with shared libraries.

The ProcessAborted exception was not exported. This causes there to be
distinct objects type of the same name to exist in different
compilation units. Therefore catches of this exception do not
correctly occur across compilation units. Specifically, the
ProcessObject is not catching the exception causing to AbortEvent to
occur.

Change-Id: Ie074699de15be7461e41c8e52a75bb0b710c576a