ENH: Only build tests for modules explicitly enabled.
If BUILD_DEFAULT_MODULES is OFF only the tests for the Modules explicitly
enabled are build to reduce build time.

Tests for modules requested by the ITKGroup_ variable are not enabled because
even for the Core Group, the test dependencies greatly increase the size of
the dependency tree.

For similar reasons, by default tests for only modules that are explicitly
enabled are built.  An option ITK_BUILD_ALL_MODULES_FOR_TESTS is added that,
when turned on, will also build all the tests of dependencies.

Change-Id: If02bde13f7e3cc47604a63ae7eec7bcfed071e5a