BUG: Fixed Affine test 32bits. Precision checks.
The testValue, testVector and testMatrix routines are now modified
to take advantage for the method:

         itk::Math::FloatAlmostEqual( v1, v2, maxUlps )

The parameter maxUlps has been adjusted to the level of precision required for
every intermediate test, and to the lower bound that still lets the tests pass
on both 64bits and 32bits. The default value of maxUlps has been set to 4, to
make it consistent with the FloatAlmostEqual() implementation default value.

This has only been tested on Linux 64/32. Additional fine tunning might be needed
for other platforms.

Change-Id: Ie4adfa0ae1d5eae3bcdf75d0e41f0196f6898cfa