COMP: Suppress Coverity defect when Examples create an ITK object
Nearly all of the ITK Examples trigger "Uncaught exception" Coverity
defects, as ObjectFactory can throw an exception if ITK_AUTOLOAD_PATH is
set and StrictVersionChecking is enabled. It is actually desired that the
user receive an exception in this case, so a code annotation should cause
Coverity to automatically mark any new issues of this sort as
"Intentional".

See: https://scan.coverity.com/models#codeannotate

Existing issues (there are 280+ currently) will have to be manually marked
or suppressed by a Coverity code model.

Change-Id: Iee8ece049caeed60e33c426515732b4c455423bb