ENH: Add scripts for automated updating of ThirdParty libraries
This adds UpdateThirdPartyFromUpstream.sh, a script to provide automated
and reliable updating of local ITK version of ThirdParty libraries from
upstream versions.

The automated update process only works for ThirdParty module code that
is stored using the Git subtree workflow. Ideally, ThirdParty modules will
be converted to the subtree workflow and use this script when they are
next updated.

This commit adds an update configuration for KWSys. Other libraries will
be added in the future.

Much thanks to Brad King for developing the initial update script and
introducing the subtree workflow.

Change-Id: Ib43fe431bc7bc7970b1f29cc5289f264aca7b5f4