BUG: Coverity 1081019: Improper use of negative value in StringTools
It's logically impossible for find_last_not_of() to return string::npos in
this function, so the Coverity warning is strictly a false positive.
However, relying on this assumption makes the code less maintainable.

Reimplemented Trim() to simply call TrimLeft and TrimRight in succession.
This reduces code duplication and does not increase the computational
complexity of the function.

Updated the test to test Trim() more thoroughly.

Change-Id: Ie62056f50ceac7362d225c98da59368eab2ea944