STYLE: Coverity 1081140: Dereference before null check
The NULLness of m_NiftiImage is checked immediately after it is set. None
of the subsequent statements or calls to nifti_read_subregion_image() or
nifti_image_load() (which both pass m_NiftiImage by-value) change the
value of m_NiftiImage. Thus, the removed NULLness check of m_NiftiImage
was superfluous.

Change-Id: I8d66cfe60c6e9adcf103f32ff2fa3e271b4b77b1