STYLE: Removed itkExtras folder
I think this folder is some leftover. There was more stuff before in this folder but it has been stripped down to the remaining unique __init__.py file. It's not good practice to have a lot of python code in the __init__.py files (especially classes and functions ...).

I propose we remove the folder if it's not used. The code in __init__.py was moved to a new itkExtras.py file and I removed the last lines because those were only there to import all the modules in the itkExtras folder.