STYLE: 4 space indentation for python files
I plan to go through the python files to do some style cleanup.
A good start is to make the code pep8 compliant http://www.python.org/dev/peps/pep-0008/

For this I will use the pep8 tool : https://pypi.python.org/pypi/pep8

To start with, I just cleaned up the indentation of the files. The indentation was not consistent throughout the files (2, 3, and 4 spaces).
I thought it would be better to clean up the indentation first, because if I do the full pep8 cleanup at once the diffs will be difficult to read.

Once this is merged I will submit a second patch with more style cleanups.