BUG: itkLoggerThreadWrapper test was disabled.
These changes restores the itkLoggerThreadWrapper test in the CMakeLists.txt
file, and adds to it a couple of calls to raise the code coverage of the class.

As a result of analyzing the logic, the operations type FLUSH was removed,
since there was no path to ever set it.

The error management logic of the pInfo data structure in the threaded function
was inverted, leaving the cases of null pointers to fall back in the default
return of the threaded function.

Change-Id: I876e10c280a8ca4569105c382f327ad601dd6e91