BUG: TestingExtractSliceImageFilter lacked test.
In this patch we add a unit test for the itkTestingExtractSliceImageFilter
class.

NOTE: The unit test uncovered a bug in SetExtractionRegion. The
dimension of the output size and index were being exceeded, causing a
crash on Windows systems.

This class is critical, given that it is used as part of the testing
infrastructure for the rest of the toolkit.

Change-Id: I0b5a80f88f22ccd754c1a6f433a2b0a57475f69a