BUG: Add support for system libtiff 4.0.0-4.0.2
This patch provides a work around for using a system libtiff for
version 4.0.0-4.0.2. The TIFFField data structure was made opaque, but
the necessary methods to access the member to inspect the tiff fields
were only made available in 4.0.3. This patch inspects the headers for
this case, and provides a copy of the declaration for the internal
structure defined in tif_dir.h for uniform access.

Change-Id: I636b7322f02c99b463db83abac6b6f52c4b9dfb3