BUG: Fix LabelStatistics and LabelOverlap to require same image size
Prior, setting images which different sizes resulted with iterators
invalid regions set causing undefined behavior, asserts, and potential
access violations.

This patch removes the overridden GenerateInputRequestedRegion method,
for the default. The default implementation ensures that all inputs
have the same requested region matching that of the first input, which
matches the output region, which is explicitly set to the largest
possible region of the output in EnlargeOutputRequestedRegion, which
is the size of the first input. Therefore if the second image is
smaller that the first, an exception will be generated in the second
input's PropagateRequestedRegion method indicating the request is too
big.

This case is testing in SimpleITK's generated tests.

Change-Id: If223fe9be1e9e1f6dc1a8459b039028840bc4910