BUG:  Ignored the case for exceeding iteration limit.
After optimization at each level we need to reset the
lower and upper time bounds to [0,1] to span the full
time range for describing the transform between the
fixed and moving image.  Prior to this commit, this
reset would only occur if the convergence value was
exceeded leaving out the case where the number of
iterations were exceeded.

Change-Id: I142ecc8a22610c79cfed37668e0f8b33b621bb82