BUG: Fix precision issues with poly-line FlatStructuring element
Changed the computation of 2D poly structuring elements to
force use of double precision. Casting issues appear to have caused
the differences between platforms.

Made changes suggested by Hans.

I suspect latest round of errors were caused by differences in threading.
These test images are tiny, so each thread ends up processing only
a couple of raster lines, which probably breaks some assumptions. This
patch sets the internal use of the dilate filter in the FlatStructuringElement
to use only 1 thread when constructing the "neighborhood". This isn't
used in any way except to get an image of the structuring element.

The longer term "solution" is to use periodic lines instead of
bresenham lines. The current implementation based on bresenham lines
isn't wrong, but the lines are not translation invariant and the
approximations of circles aren't as good as with periodic lines.

Change-Id: Ib7526f009792f4ed34dc1c03195ccedf3a4ec446