COMP: Fix checks for system libtiff
The wrong cmake include path variable was used.

CMake's check_symbol_exits requires the result to be fully linked,
which is difficult for libtiff due to possible additional
dependencies on zlib and jpeg. By using check_type_size on the
function name (a pointer), the linking dependency is removed.
The symbol exists in tiff 4.0.3, and requires library.

Change-Id: I46b46af7ee58fcb969e2ba213c542ea1efa955ce