COMP: Remove clrLine in itkExtras
A special escape character was used in python to display the loading process.
This was intended to display the loading sequence on one line in the terminal.

When there are errors in the python tests, the error messages are cluttered with : [2000D[NON-XML-CHAR-0x1B] in Cdash
Anyway this hack also seems not to work  on OS X's terminal.
I suppose we don't want to support all types of terminal types and configurations here.

Change-Id: I387bab704718534dcdc9305944894d7de383c236