COMP: Fix the python import and progress callbacks
I put back the clrLine code removed in my previous commit.
In fact the bug in CDash were the ascii string was appaering was due to a bad comparison
in the auto_progress function. This was due to the changes done during the pep8 clean-up.
Even if you pass 2 as progressType, "if 2" will return True. The python tests used progressType=2,
so the wrong callback was used.

This commit reverts the changes from the previous commit, fixes the bug for progressType=2,
and adds some explanations for the clrLine signification.

Change-Id: Ibb26397650b394882ef52877181b116535cb064d