ENH: Add setters for the overlay functor in labelmap overlay filters.
This will allow a user to draw label overlays or contours with a user-
supplied colormap, if required, while preserving existing behaviour
for those users happy with the defaults.

This also fixes a (previously unimportant) bug in comparison of
LabelOverlayFunctor and LabelToRGBFunctor classes, and fixes an
uninitialized m_Opacity member in LabelOverlayFunctor.

Also added unit tests for custom overlay and contour colormaps.

Change-Id: Ic5d783d8397abb0f652e8dd9a90c44f39d07a08c