STYLE: Remove unused compile all code for Python Wrapping
This code is commented out since very long time, so not tested with recent python versions.
Anyway the compileall.py file is now a module by itself so this would need some fixes.
If it's not used and not tested, we should just remove it.

If we need to byte-compile the python files one day for whatever reason,
writing some lines to do it from scratch should not be too difficult.

Change-Id: I876213993ecee932ee87d1ef98618cf8b7407aa9