STYLE: Add ITK_NULLPTR supporting c++11 checks
In c++11 there is an explicit nullptr type that introduces a new keyword to
serve as a distinguished null pointer constant: nullptr. It is of type
nullptr_t, which is implicitly convertible and comparable to any pointer type
or pointer-to-member type. It is not implicitly convertible or comparable to
integral types, except for bool.

This patch adds a define for ITK_NULLPTR as nullptr when c++11 is
available.

clang-modernize -p ITK-newclang/ -include ITK/Modules/ -exclude ITK/Modules/ThirdParty -use-nullptr

Change-Id: I2f9aeb7d92581cc84a75bdc6a00b79169e72100a