DOC: Corrected documentation for threshold boundaries.
Previously, the documentation for itkThresholdImageFilter was
incorrect for the ThresholdAbove, ThresholdBelow, and ThresholdBetween
options because they incorrectly stated the handling of pixels in the
image equal to the threshold value itself.  The previous comments
stated that pixels equal to the threshold value will be changed to the
outsideValue.  But according to the code, the true handling of pixels
equal to the threshold value is that they are not affected by the
threshold.  The documentation has now been corrected.

The corresponding test was enhanced to add explicit testing of the
threshold value, and the new tests pass, further indicating that the
new documentation is correct.  The test was also extended to handle
cases where the inputValue is not 0 for generality and to make the
code clearer.

Change-Id: I3fc7f36a532e7b64b03d9591e31057e7d1cc745d