ENH: Improve Resample's use of pipeline inputs
Make the input transform to the ResampleImageFilter a proper pipeline
input. The private ivar m_Transform can safely be removed. It's is now
a named required input to the process object. It's is automatically
verified that it's non-null before pipeline execution, and the MTime
is automatically propagated during pipeline execution. NOTE: There is
a subtle change in the behavior of ResampleImageFilter::GetMTime.

The ReferenceImage is not a named input, remaining optional an on
input 2.

SmartPointers were changed to raw pointers for object being held by
the pipeline.

Change-Id: Id3a67bef2646bd4fbbe4391e4c7b8da8a2194895