ENH: Improve DataObjectDecorator with Modifiable, Graft ReleaseData
Allow the component held by the DataObjectDecorator to be
modified via the GetModifiable method. The MTime of the held component is propagate to the
decorator.

Additionally a Graft method is added which just copies the pointer so
the two object refer to the same component.

The Initialize method was implement to add support for ReleaseData
during pipeline execution.

Because the ImageRegistrationMethod uses the output as input for the
initial transform, the output data object can't be released before
update. The image filters also by default operator by no releasing the
output by default.

Change-Id: I2bea47efdcfdadcfa6a52c7e650435e63db34d46