ENH: Use Transform base class as default template parameter
With the ability to set an initial transform, it is no longer
necessary for the ImageRegistrationMethodv4 class to directly
construct the output transform. This type can be copied/cloned/grafted
from the InitialTransform.

This reduces the permutations for methods which need to be
instantiated making it possible to wrap this class for SimpleITK and
WrapITK.

Change-Id: I4a3af0aad48bbf3203d7b3f1138547d04cdde684