ENH: Move the NumberOfIterations to the Optimizersv4 base class
In Optimizersv4 directory, either MaximumNumberOfIterations or
NumberOfIterations were used to reflect the upper limit for the number
of iterations. However, in this patch a consistent name is used in all
classes as "NumberOfIterations".

Also, this variable is moved to the "ObjectToObjectOptimizerBase" class,
and all other optimizers inherits that from the base class.

Change-Id: I03f55943b52b5cfe59d86557667621f4ad4c10a7