BUG: Buffer not null terminated
Coverity reports:
buffer_size_warning: Calling strncpy with a maximum size argument of
32 bytes on destination array "header.filename" of size 32 bytes might
leave the destination string unterminated.

If the buffer is treated as a null terminated string in later
operations, a buffer overflow or over-read may occur.  In
itk::​BioRadImageIO::​Write(void const*): The string buffer may not
have a null terminator if the source string's length is equal to the
buffer size.

Change-Id: I3b2a5fa83b0354580a99a48907a0518e785f2a6e