ENH: Use TransformParametersAdaporBase on Transform base class
The TransformParametersAdaptor class and
TransformParametersAdaptorBase class were redundant. By making the
Base version operate on the base Transform class it now provides a
generic interface to the ParameterAdaptors independent of a specific
transform.

This enables transform adapters to be used with the
ImageRegistrationMethodv4 without the an explicit template parameter
for the transform type.

Change-Id: Ie7b4fc8ad859d92ddc3e1864b9066ae357f3278e