BUG: Exclude Remote Modules the default ON Group values.
When a Remote Module is explicitly enabled, with

-DModule_<remote-name>:BOOL=ON

it will initially download the sources into the ITK source tree.

If a fresh build has been created, and the module was previously enabled,
it will be automatically enabled. In this patch, we ensure that all
Remote Modules are OFF during initial CMake configuration to generate a
consistent, predictable user experience.

We also fix the indexing of modules for Doxygen. Bridge modules are now
visible in Doxygen, and Remote Modules, that do not have "ITK" in their
module
name, are added to the appropriate Doxygen Group.

Change-Id: I38cecaec345fb7f17d1812682c5e69934563d3d6