ENH: support learning rate estimation for regular step optimizer
This optimizer has a fixed (regular step size) independent of the
gradient yet dependent on the scales. The super class method
EstimateScales method is over-ridden to account for the gradient
independent scale of the learning rate.

To demonstrate the scales estimator, an additional flag has been added
to the ImageRegistrtion1 test.

Change-Id: I25561344e4edda8abf85c29d7f183f192849e12d