BUG: GDCM does not apply rescale slope / intercept on write.
Make use of the InternalComponentType, which is the pixel type
before rescale slope/intercept have been applied when reading from disk.

This internal component and rescale slope from the reader can be used in
a writer for the DICOM image with

  dicomWriter->SetImageIO( readersGDCMImageIO );
  dicomWriter->UseInputMetaDataDictionaryOff();

Make sure the written DICOM images are verified against baselines in the
tests.

Change-Id: I1cecec83d2bfeac5a8e64b77062a46dcd49746f7