ENH: Convert seven ImageRegistration examples to ITKv4
Seven registration examples are refactored to use ITKv4 registration
framework. Also, the explanations related to the deprecated
features are remove, and new explanations are added to reflect the
new features.
New examples related to the ITKv4 registration are set in a different
directory than ITKv3, and any conflict is avoided since the binaries
of ITKv3 examples are created using an "ITKv3" prefix to avoid any conflict.

Change-Id: I9b9f0d5e65eedfd38b92ba4504f45238d7c95ad5