ENH: New test for itkLabelShapeOpeningImageFilter
The /Modules/Filtering/LabelMap/test/CMakeLists.txt has been modified so
that the existing test itkLabelShapeOpeningImageFilterTest1.cxx is
included, and the itkLabelShapeOpeningImageFilter class is tested. The
baseline image generated when running the test locally is being committed
along with this patch. The test compiles fine under Visual Studio 2010,
and runs successfully in Windows 7 64-bit.

Change-Id: I815cb6f8ddf1859b91b15afb9e579dad61c13f1d