PERF: Remove large foot print of PDF derivatives.
ITK-3286 Describes that the memory footprint is very
large for many threads and MattesMutualInformation
due to the amount of memory being allocated per thread.

Allocating memory is also the most time consuming portion
of the MI algorithm and drastically restricts the
scalability of the algorithm to many threads.

Performance Enhancments for minimizing number of
memory bufers needed to avoid false sharing by using
a queue based accumulation per thread that is periodically
syncronized with the primary pdf.  This approach
allows the time consuming part of each thread to
occur and the syncronized writing to the JointPDFDerivative
in parallel.

Each computation of a point was previously allocating new
memory.  This patch set checks if the currently allocated
memory is the proper size, and then re-initalizes that
memory if possible.

Change-Id: I72f1d1ea2be71a883a47fe71afc2724e5ec88fb8