BUG: Set output type to unsigned char when writing uchar in GDCMIO test.
In itkGDCMSeriesReadImageWriteTest, the GDCMImageIO is re-used on a writer
that is writing a unsigned char image.  However the ImageIO was originally
used to read an unsigned short image, and writing unsigned char data to an
unsigned short DICOM file is not supported by DICOM.

Use SetInternalComponentType to set the correct internal component type when
writing the unsigned char file.

Change-Id: I5410afc22678990a173e3bcfe578cfc984e16763