BUG: Corrected processing of last sigma
The m_NumberOfSigmaSteps should reflect the true number of sigmas that are computed.
In the itkMultiScaleHessianBasedMeasureImageFilterTest, the
m_NumberOfSigmaSteps is set to 10, the lower sigma is 5, and the upper is 10.
But if the sigma values are printed out in
itkMultiScaleHessianBasedMeasureImageFilter.hxx while running the
test, only 9 sigma values are printed as follows:
5.000000 5.400299 5.832645 6.299605 6.803950 7.348672 7.937005 8.572440 9.258747

Clearly, the last sigma value is missing.  If the sigma value after
the loop is printed, it is 10.000000 with a non-zero number in a much
later digit.

This patch corrects the operation of the filter by changing the loop
and cleaning up some of the variables.
It also adds a regression image test since previously
there was none.

Change-Id: Iecdd472e5a990b5664f8db51c4732c744b08b33e