BUG: Fix overflows computing size of read tiff image
When multiplying the height and width of an tiff image (32-bits),
at least one element should be converted to size_t to help prevent
numeric overflow.

This change enables reading of larger RGB and large 2D image, that
were previously truncated when reading. This does not address
potential buffer overflow or resource limits which could occur.

Change-Id: I192370606677c8dc1a903c67c2a0701026256d5a