BUG: gdcm::StringFilter recognizes backslash delimiter.
For the basic numerical types, gdcm::StringFilter::FromString was not
using \ as a delimiter for multiple value elements.  Skip \ during
istringstream extraction as per upstream

  https://sourceforge.net/p/gdcm/gdcm/ci/61f39a3/

UT (Unlimited Text) is removed from FromStringFilterCase because it is already
handled by if( VR::IsASCII( vr ) ) { return s; } (gdcmStringFilter.cxx 296).
For more details, refer to "gdcmVR.h" where all "character string" types are
defined using the macro TYPETOENCODING.

This fixes uninitialized writes as detected by valgrind and the elements were
verified to be corrected populated with the correct values in the populated
tests.

Change-Id: If34d5fa9dd532949d309d7a8ba501cf508b33e53