BUG: Remove Zeiss 2-channel support code in TIFFImageIO ( and LSMImageIO )
There are no tests for this branch of code. It only works in 3D under
certain cases. There are apparent bugs in the code such as not setting
all of the input image, and logically dead branches. Some downloaded
test images for similar described format do not load. This may effect
files with the extensions tif, tiff, or lsm.

Removal of this code will enable expansion of the current code to more
generically support multi-sample per pixel images.

For more robust reading of this type of image the SCIFIO remote
module should be used.

Change-Id: Idb98d69538db3bc8520881ea4a5e22476238c41e