ENH: Refactor duplicated code to read a page
Created a method to read a single page or sub-image in a tiff
file. Simular code was both in the ReadVolume method for 3-d and the
Read method for 2-d. The code block from ReadVolume was used because
it contained more typed cases. There for this patch does improve the
2d reading capabilities for tiffio, so that they are consistent with
the 3d.

Change-Id: I0fbfb19f04435a703bb3fa09ea4e79179592e443