BUG: Do not perform itk_download_attempt_check when not building ITK.
When processing WrapITKConfig.cmake, for example with a find_package(WrapITK),
it will include ConfigureWrapping.cmake, which in turn includes

Wrapping/Generators/GccXML/CMakeLists.txt
Wrapping/Generators/SwigInterface/CMakeLists.txt

In this case, the itk_download_attempt_check macro is not defined, but the
check is not required.  Only perform the download check when building ITK,
which is detected by the ITK_BINARY_DIR variable, which is generated by

  project(ITK)

in the top level CMakeLists.txt.

ITK-3319

Change-Id: I372357eefc48abe783b5fb59a1a55d3b3a050cfb