ENH: Refactor per pixel conversion function to per scan-line method
This patch changes the implementation of conversion from the TIFF
scanline buffer from a function call on each pixel to a function call
per scanline. The enable efficient copying of similar scanline and
better reuse of variables for palette look ups.

This can increase the performance of reading grayscale images by up to
3X for cached files.

Change-Id: I24bb744fda35629f1c1ed7154e26e5708e912059