COMP: itk::Command-derived class used pointless dynamic_cast.
There is a common pattern in the ITK Examples and tests:
Use dynamic_cast to downcast the 'object' parameter to
the 'Execute' method, then ignore the return value.

Given that itk::Command-derived classes are tuned to a
single datatype, unless a programmer has screwed up,
dynamic_cast will always succeed. This means it might as well
be static_cast, especially if the return value is never checked.

So I replaced the dynamic_cast with static_cast where they
occurred in the Execute method of itk::Command derived classes
with static_cast.

Change-Id: I77f0735598ad5657e2a7a6114fee004ab1daf3d9