BUG: Expand a test to reflect multi-metric registration bug
In this patch the "itkSimpleImageRegistrationTest2" is extented to use
an additional rigid registration stage with a multi-metric that contains
only "one" metric component.
This test is supposed to show a bug in ITKv4 registration framework when
it is used with multi metrics.
This test will be failed by returning an exception because current
ImageRegistrationMethodv4 always assumes that multi metrics have more
than one component when it tries to set the metric sample points.
Also, in the extented test, the multi-stage registration is implemented
in a cascaded way that is different than the methods used in other
existent tests, and it can be a bonus itself.

Change-Id: Ibfa9b5c65794ec5cf369c9aef84bb822eca799f0