BUG: Fix wrapping .i, .idx CMake dependencies.
Each swig interface .i generation depends on the .idx files, which are
populated with the template parameters for the classes defined in that .i
file.

Every ITK Module gets a .i file, but with every wrapped module, there are
"submodules" that correspond to the ".wrap" files for the module. All the .idx
files for each submodule are listed in the .mdx file for the module.

Make sure that all the .idx files for the current modules are added as a
dependency for the .i being built.  Previously, the ${idx_files} was only
partially populated.

This fixes the parallel build with the CMake Ninja generator.

Change-Id: I95350d724c43896756cd83dca09d398d9d037fcf