ENH: LandmarkBasedTransformInitializer supports BSplineTransformType
Only the following transform types were supported by the
LandmarkBasedTransformInitializer class:
  VersorRigid3DTransform
  Rigid2DTransform
  AffineTransform

Now, support for the BSplineTransform is also added.
Also, I moved this class from the Modules/Core/Transform/ to the
Modules/Registration/Common/ to avoid module dependency cycle.

New implementations needed the ITKImageGrid be added to the
dependencies, and if this class was kept in ITKTransform, it would cause
a module dependency cycle.

Change-Id: I09b895627fcace1915ea8ade76f0f7f7273b5ffa