STYLE: Some minor code cleanup in the wrapping setup
The WRAPPER_DO_NOT_CREATE_CXX variable is never used so it was removed.
The macro itk_end_wrap_submodule() is in this case no more needed; instead
call directly the itk_end_wrap_submodule_all_generators() macro.

Remove mention of WRITE_WRAP_CXX in comments; this was an old macro which no more exists,
the writing of the wrap_.cxx.in file is now done in the gccxml generator.

Simplify code for filling up the wrap_.cxx.in file. Intermediate variables were previously used, which is not needed.
Remove @CONFIG_WRAPPER_EXTRAS@ from wrap_.cxx.in file, we are not setting this variable.
Cleaned up the spacing in GCC_XML_TYPEDEFS and  GCC_XML_FORCE_INSTANTIATE for nicer output in .cxx file.

Change-Id: I2806589fdd21e93f08784f472aa13f79e460a1f9