ENH: Add direct initialization to SyN registration
It is important to be able to save the internal state
and restore the internal state components necessary
to accomplish a SyN registration.  This patch
adds the logic so that it is possible to restore
the state from a properly formatted initializer.

Changed name of typedef to avoid warning
COMP: itkBSplineSyNImageRegistrationMethod.hxx:328:
          warning: declaration of 'DisplacementFieldTransformType'
                   shadows a member of 'this'

Change-Id: I50b18e6d46b93d10d4b6042d43f4baede8a87d92