ENH: Add test to exercise issue with Transform::GetInverse
Hans Johnson first noticed these failures with the BRAINSTools
package after a recent change to the GetInverse function pointed
out two things:

1. There are a lot of places potentially where a single transform is
shared across threads.
2. Transform::GetInverse is not thread safe if you call it to get an
inverse and store it in a variable shared across threads.

This test exercises the problem. It fails reliably on OSX, and less
frequently on Linux. Behavior on Windows is not tested.

Change-Id: Ic9f5f62422e9529ab4ce4928dfbaf284cd44288f