BUG: MatrixOffsetTransformBase::GetFixedParameters not thread safe.
The call te this->m_FixedParameters.SetSize(NInputDimensions) is not thread
safe.  For example, it can cause a crash when GetInverse is called on the
Transform.  Remove this call, since it already occurs in the constructor and
NInputDimensions is a template parameter.

Also remove the trivial call by some classes that inherit from
MatrixOffsetTransformBase.

Co-authored-by: Kent Williams <norman-k-williams@uiowa.edu>

ITK-3224

Change-Id: Id6c62404c2b579ba96497f4466b6b84ba3967d54