PERF: Don't recompute InternalInverse repeatedly
SpatialObject supposedly caches an InternalInverse transform
that's used by SpatialObject::IsInside to test if points
are inside the SpatialObject.  But a 'convenience' function
SetInternalInverseTransformToWorldToIndexTransform is called
at the top of IsInside in each of the derived classes, and
it was causing the inverse of the IndexToWorld transform to
be recomputed every time IsInside was called.

This patch just moves the GetInverse to where the IndexToWorld
transform is actually changed.

Change-Id: I616246263c1c0ba58dbef6ab0713fdbb2d0f05eb