ENH:  Adding sparse points to B-spline field estimate.
Originally, the itkDisplacementFieldToBSpline filter took
a displacement field represented as an image an estimated
a smooth version based on b-splines.  This filter underlies
a couple of the registraiton methods/transforms involving
B-splines and so we need to modify this filter to also
accept point sets and specified B-spline domains (even in
the absence of a specified dense displacement field).

Change-Id: I2e6f848e128ada65fb20ca574ccfc7d0fe1cd81b