ENH: Query libtiff to determine compression support
Depending on the version of tiff used and how it is configured will
yield support for different compression codecs. The libtiff method
TIFFIsCODECConfigured is now queried to determine support.

If the compression is not supported than both the scanline and the
RGBAImage read methods will fail. The codec is now checked and if it's
not supported than a better exception is thrown.

The TIFF Compression RGB baselines image have been corrected from RGBA
images to RGB image.

Change-Id: I493cbb9d602f7902ced0c92f664c69238bf4bb52