ENH: Expand interface for selecting threadpool
Enabling the threadpool could be done with an
environmenal variable.  This provides a the
API for setting the threadpool on programatically.

If the API is used to use the threadpool, that will
take precedence over the environmental variable.  The
environmental variable is only used if the developer
does not explicitly request the threadpool.

Change-Id: Ibd035f1d6c4de41cd883266a30e9f4fb9e4c3e69