BUG: Fix segfault when setting displacement field as null
This patch addresses a segmentation fault we setting the displacement
field image to a null pointer. The initial state of the displacement
fields is with a null displacement field image. This patch explicitly
sets the fixed parameters to 0.0 when the field is a null
ptr. Additional checks are added to only set the interpolator's image
when both the interpolator and the displacement field is set

Change-Id: I35202a0049f5ce0921690fca94901f48b639785c