BUG: restore the center point fixed parameters for scale transforms
The Scale Transforms support a center point around where the
transformation should occur. However these parameters were not being
reported by the FixedParameter accessors. This class was modified to
fully support the superclass variables including center and fixed
parameters, and compute offset is also called to ensure consistency
in the parent classes state and the children.

Change-Id: Ie7482c3d3395cb5cf18115c68749d1bb4f4c8071