BUG: Rename Module to IOFDF
The modules name is inconsistent and causing not to be used as a
module in ITK. Previously ITK was calling itk FDFImageIO, while
internally it was called ITKFDFImageIO. Neither name followed ITK
conventions. The correct name for an external module does not begin
with the ITK prefix and for this case should be: IOFDF

The name of the module and related variable, libraries, export
specifications etc, have been rename consistently.